examples/*/package.json: Add "start" targets
Currently, start:browser starts everything in development mode.  I think
it would make more sense if it only started the browser part (including
the development web server) and if the start target was used to start
the whole thing.  "npm start" is a standard top-level npm command which
invokes the "start" script, so people are used to it.

I added some blank lines between logical groups to help readability a
bit.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>