examples/*/package.json: Use prepare instead of prepublish
If we want the build to run automatically after an "npm install", we
should be using "prepare" instead of "prepublish", as described in [1].

I changed the command in the electron version to do a clean, just like
the browser version does.

I tried to keep the steps in separate, logical groups (build, run,
watch, test, etc).

[1] https://docs.npmjs.com/misc/scripts#prepublish-and-prepare

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>