file-dependency-updater: use awaitWriteFinish
When tsc writes a file, it first truncates it (open with O_TRUNC) and
then writes to it.  At the inotify level, it generates two modify
events.  Chokidar throttles the change events to max one per 50 ms, so
we only receive the first one, that corresponds to the truncation of the
file.  If the file-dependency-updater script is fast enough, it will
make a copy of the file before tsc had the chance to write it.  Since
the second inotify event is suppressed, we'll end up with an empty file
in the example's node_modules directory.

The solution favored by chokidar seems to be to use awaitWriteFinish
[1], which acts like a "debouncer".  It waits until the file is stable
for a certain time before reporting the change event.  I chose a delay
of 100ms.  Actually, I think that anything over 50 ms would be safe: if
for some reason there's more than 100ms between the open/truncate and
the write, we will get an event while the file is empty.  However chokidar
chokidar won't suppress the second one, because more than 50ms separate
this one from the first one.  So we might copy an empty file in the
first time, but we will do another copy after the file will have been
written.

I also noticed we could listen to "raw" events.  In that case, chokidar
gives you all the events that the underlying system sends.  But I think
it's better to have just one event once everything has settled down,
instead of multiple events really quickly.  There's more chance of
triggering race conditions at the next level (e.g. webpack) if we copy
empty files and then copy the real files.  Also, the events differ per
OS, and we don't want to deal with that (that's the point of chokidar).

[1] https://github.com/paulmillr/chokidar/issues/365

Fixes #125

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>