Add fontend debug configuration for vscode
This adds the Launch Frontend debug configuration to vscode.
This is used to debug the browser example's frontend.

Note that you need to to start the dev server before you can launch this
config using npm run start:frontend

It also renames the previous Launch App config to Launch Backend.

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>