Update travis config to add the ui tests to the ci
In order to do so, the ui test npm start script was modified to start the server before running the tests. Because the server start and the tests are running asynchronously, the tests are waiting for a valid connection to be up on the host:port of the test server.

Signed-off-by: guy perron <guy.perron@ericsson.com>