Delete only once process from process manager
Before this patch the process id could be deleted twice from the process
manager.

A call to process.dispose() with killed === false would delete the id and
then an exit callback was called with a call to process.dispose().

This patch deletes the id only once the process is killed.

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>