Refactor Process so that it doesn't need to be disposable
Care must be taken to dipose of event handlers attached to its event
emitters however.

The right way to delete a term is to delete it from the process manager,
this will kill the process and send a onDelete event with the terminal id.

Owners of event handlers on a terminal should listen to that onDelete and
dispose of their event handlers like is done in base-terminal-server.

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>