Use proper mocking for the storage, widget-manager, problem-manager tests
This patch reverts d68a165b5f18199a79d75ddbc29aa1448b6cf78b ([logger]
extracted `AbstractLogger`) so that the logger code is unaffected by the
test requirements.

It also refactors the storage, widget-manger and problem-manger tests so
that they use inversify dependency injection to inject the mocked
dependencies.

The strategy used by this patch for testing is that:

 - An mock implementation of the depedency is provided, in this case the
   MockLogger.

 - This mock class can be customized according to the needs of the
   particular test using sinon.

 - Inversify Dependency Injection is used to inject the mocked
   dependencies rather than directly using the constructor.


Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>