Don't read subdirectories contents while reading a directory
Before this patch when requesting the contents of a directory with depth
0, we also read the contents of the subdirectories.

This patch makes it so that listing a directory with depth 0 will only
read this directory contents.

Also, this patch removes the hasChildren property since this is no longer
used. And replaces it with a function called mayHaveChildren that will
return true if a URI can have children.

This patch also passes the overwrite options as the clobber options to mv
so that the overwrite behavior is respected.

Fixes: #646

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>