Refs #532:Add passthrough keybindings, make ctrl-c work in terminal
Currently, typing ctrl-c in the terminal does not yield the expected
result, which is sending a SIGINT to the running process.  The problem
is that a Theia keybinding is registered for ctrl-c (to copy).  Theia
catches it and thinks you want to copy.

When the focus is on a terminal, we don't want Theia to handle the event
itself, but rather let the event propagate.  xterm.js' keyboard event
handler will catch it, and things will work as usual.  To that end,
this patch introduces the concept of passthrough keybindings.  When
the user invokes such a keybinding, Theia does nothing and lets the
event propagate.  To specify a passthrough keybinding, one needs to use
"passthrough" as the command id, which is treated in a special way by
the KeybindingRegistry.

However, we also need the concept of event priority, to make sure that
when we are in a terminal context, that keybinding will take precedence
over the global one.  As a simple approximation, I made the
KeybindingRegistry consider bindings that specify a context before those
that don't.

The terminal module registers passthrough commands for all the key
combination xterm.js recognizes right now:

  https://github.com/xtermjs/xterm.js/blob/v3/src/Terminal.ts#L1684

Some improvements should be made for Mac though.  To match xterm.js'
behavior, the alt-<key> combinations should only be registered on
non-Mac computers, but I don't know how to do that.  Also, I suppose
that on Mac, the ctrl-<key> bindings should be registered with the M4
modifier?  Finally, there's the cmd-A shortcut registered specifically
for Mac.  It would be nice if somebody that has access to a Mac could
improve that.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>