Add MultiRingBuffer class
The MuliRingBuffer is a ring buffer that allows multiple readers to read
from a single ring buffer and maintain their own reading positions.

This class will be used by the terminal in later patches to contain the
terminal output history and share it why any terminal widget attached to
this terminal.

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>