Fix terminal restoration for shadowed tabs
The div used to open xterm.js needs to be visible when the xterm.js open
function is called so the size of the characters can be calculated using
getBoundingClientRect()

This patch makes it so that open is only called when the terminal is
visible.

This patch also saves/restores the terminal title so that the restoration
is smooth and makes sure that the resize callback is not called on a term
that was disposed.

Fixes: #818

Signed-off-by: Antoine Tremblay <antoine.tremblay@ericsson.com>