Fix default label provider long name
It currently returns the parent (directory) name, but it should return
the element name.

When trying to add a test for label-provider.ts, it tripped on the
import of the css file, trying to execute it as typescript/javascript.
I moved the import with the other css files in
frontend-application-module.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>