[core/keymaps] Refactored keybindings and user keymap parsing now checks for collisions
Fixes #1139; Fixes the keymaps.json validation and adds tests for it.

Before, when registering keybindings, the whole scope was cleared when a collision occured.
This commit only drops the last colliding binding.

The keybindings have been refactored a bit, for the sake of not duplicating the code in multiple functions.

The keymaps.json file validation is also tested now, as the schema before was wrong and triggered a lot of false positives.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>