Allow changing log levels of different loggers independently
As we add more logging, it will be useful to split it into different
loggers, so that it's possible to only enable the verbose logging of one
particular component of Theia.  This patch allows the user to set the
verbosity of each logger independently.

First, it adds names to loggers, so that we can refer to them by name.
The loggers are then uniquely identified by these names.  This allows
the frontend and backend to both create loggers of the same name, which
will refer in the end to the same logger.  If the user changes the log
level of a logger, it will affect both the frontend and the backend.
Using names as the unique identifier of the loggers makes the numerical
id obsolete, so it is removed.

Log levels are specified using a config file passed to the new
--log-config option to the backend.  The expected format is:

    {
      "defaultLevel": "debug",
      "levels": {
        "languages": "trace",
        "task": "info"
      }
    }

"defaultLevel" sets the log level to use for loggers not otherwise
specified.  "levels" should contain an object where keys are logger
names and values are log levels.  Unfortunately, it's not possible to
validate logger names, as the list of existing loggers is not known at
startup (and a new logger could appear at any time).

The new option --log-config is mutually exclusive with --log-level,
to keep things simple and unambiguous.  --log-level=foo behaves the same
as providing this config:

    {
        "defaultLevel": "foo"
    }

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>