Reorganize the tslint.json files
The end goal of this is to have a tslint.json at the root of the
project.  By putting it there, tslint will automatically pick it up, no
need to explicitly point to it with -c/--config.

This will help when we'll have tslint integrated into Theia as a
language server.  It will will automatically find the rules without any
extra configuration.  Eventually, we'll want to offer the user a
preference to point to a specific tslint.json, but that will come later.
And anyway, it's better if it works out of the box, without having to
set a preference.

I am a bit confused by the current naming (strict.tsconfig.json and
tsconfig.json).  One contains rules we want as warnings, and the other
the rules we want as errors, so I renamed them warnings.tslint.json and
errors.tslint.json.  The tslint.json at the root just imports both
files.

Apparently, all.tslint.json is no longer used, so I removed it too.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>