RipgrepSearchInWorkspaceServer: reset style with --colors=X:none
This patch adds the `--colors=X:none` option, where `X` is `path`,
`line`, `column`, and `match`, to the array of arguments passed to
ripgrep(1) so as to ensure that all the styles are reset (in a known
state) before setting the specific foreground colors afterwards.

This is important since ripgrep's `--colors` option _adds_ attributes to
the current style, and ripgrep's default style could change in the
future, or the user could execute a custom version. Because we're using
specific ANSI escape sequences to actually split the results's different
parts (path, line and column numbers, and match region within the
matched line), the espace sequences that ripgrep writes must never
change.

Also removing the bold component of the expected start of match escape
sequence (`\x1b\\[1m`) because of this style reset.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>