GH-1713: Fixed keybinding issue in FF.
Same keys (such as `Minus`, `Quote`) might provide different
`event.keyCode` integer values on the same platform with different
browsers. Instead of just relying the `keyCode` value, first,
we check whether we can interpret the `event.code` string, if so,
we use that. If no, we fall back to the existing logic.
This should fix the FF issue.

Closes #1713.

Signed-off-by: Akos Kitta <kittaakos@gmail.com>