[example-browser] UI test fix
The test trying to open and check the `View > Problems` view was failing because of a timeout.

What was happening is that the problem view was already opened. So when the test clicks onto `View > Problems`, it would close the pane instead of opening it. The test would then fail, waiting for the closed pane.

This commit simply verify if the problem view already exists for some reason.

Fixes #1887

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>