Removed FrontendApplicationContribution from AbstractViewContribution
The application contribution callbacks where only used to do
`initializedLayout` which is undesired in most cases. As a result the
`FrontendApplicationContribution`wasn't bound in many cases.

This also removes the plug-in as well as extension UI from being
initially opened. the GitHistory was eagerly started, too, due to
reacting to git log events. (fixes #2140)

Signed-off-by: svenefftinge <sven.efftinge@typefox.io>