Enable tslint rules await-promise and no-return-await
We have a few occurences of problems of these types in Theia.  While
they may not constitute bugs, they may be quite misleading.  For
example, doing an "await" on a function that returns void instead of
Promise<void> can make you think that we'll wait for the function to
finish its work before continuing, when in reality it's not the case
(the function may spawn some asynchronous work and return a promise, but
doesn't currently).

Change-Id: Idee435696b9f1d639afcafc63784fc41b39b6efd
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>