Fix lerna concurrency=1 when testing on Windows
scripts/lerna.js and the "test:theia" script in package.json do not
agree on the format of the --concurrency flag.

scripts/lerna.js has "--concurrency==1", while package.json has
"--concurrency 1".  When testing on Windows, we end up with both on the
lerna command line, because scripts/lerna.js doesn't recognize that the
flag is already present.  Also, I don't think "--concurrency==1" is
valid.  "--concurrency 1" and "--concurrency=1" are valid.

Funnily enough, when you pass the flag twice, it is not applied:

$ npx lerna run test --concurrency=1 --concurrency=1
lerna info version 2.11.0
@theia/ext-scripts: yarn run v1.10.1
@theia/ext-scripts: $ echo 'skip'
@theia/ext-scripts: skip
@theia/ext-scripts: Done in 0.07s.
@theia/plugin: yarn run v1.10.1
@theia/application-package: yarn run v1.10.1
@theia/tslint: yarn run v1.10.1
@theia/plugin: $ theiaext test
@theia/tslint: $ theiaext test
@theia/application-package: $ theiaext test
...

Change both files to use "--concurrency=1".

Change-Id: I64c1cb27a603fbb951ca07cea43f44cd104d4b60
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>