Make the diff navigation header static
Fixes #3456

Updated the `git diff` navigation header to be static (fixed).
Currently when there are a lot of changes present in the files changed
list, we lose the header when scrolling. This means that the navigation and
information present in the header is lost so users must always scroll
to the top to access them. This change addresses that issue by providing a
new css class and overflowing only the files changed list to be scrollable.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>