Fix 'Remove Folder' and 'Add Folder' commands
Fixes #5070

Currently, when the `navigator-contribution` attempts to register a context
menu for the commands `ADD_FOLDER` and `REMOVE_FOLDER` the commands have
yet to be officially registered since they are waiting for the `preferences`
to be ready. This results in an error thrown when building the menu, since
the command is not registered. Instead of registering the context menu items
with other items, the logic was moved to the `postConstruct` to ensure that
the command is actually registered.

Additional Fixes:
- fix typo from `toDispoaseAddRemoveFolderActions` to `toDisposeAddRemoveFolderActions`.
- fix typo in the word `wodget` to `widget`.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>