[typescript] re-use Theia launching program for TS LS
When we launch the TypeScript Language Server, we start it using
the hardcoded executable "node". This works well in practice for
most cases. However when we have a packaged Electron application,
installed by a user, it could be that the environment does not have
"node" immediately available in its path, or not a proper version,
as needed by tsserver.

So why not reuse the node executable that was used to launch
the Theia application, to launch the LS? In the Electron case,
it will be the Electron executable bundled with the packaging.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>