Fix error cannot read property 'forEach' from scm
Fixes #5363

If the application is started without a workspace, or with a workspace without a project under
source control, the error `cannot read property 'forEach` is thrown by the `scm` package.
This is due to the fact that `this.statusBarCommands` is not instantiated unless a repository is present.
This means that when we iterate the array, the error is thrown. The minor fix is to simply instantiate it
to an empty array.

- minor: format the import statements in `scm-contribution.ts`.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>