Resize octicons to match fontawesome icons
The octicons are SVG and are by default sized 16x16 pixels.
The fontawesome icons on the other hand are text and sized
depending on the font size and line height (which results in our case
in 12.5x12.5 pixels).

This change will force the `height` and `width` of the octicons SVG tag,
displaying them the same size as the fontawesome icons.

Fixes theia-ide/theia#5534

Signed-off-by: Mathieu Agopian <mathieu@agopian.info>