Sort the extension names in the 'about' dialog
Fixes #3886

Add minor enhancement to the example `about` dialog present in the example applications
to display the list of sorted extension names. Previously the list was unsorted which
meant it was harder to identify which extensions were in fact present, and a sorted
list is more adherent with good UI practices.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>