[ci] Fix travis issue with libffmpeg test
This way ensures that despite the caching done in our CI/CD, the
replacement scripts still work, and what we pull is clean.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>

Update ffmpeg error to not fail CI

- updates the ffmpeg error in the post-install script so
that it displays a soft error instead of throwing an error
and failing the CI.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>

[ci] cleanup electron before saving cache

Make use of the "before_cache" section in .travis.yml, to remove
electron and realted files, before the cache is updated. This
makes sure that the libffmpeg test is run fresh, for CI jobs that
benefit from the cache, so we can detect eventual problems ASAP.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>