Fix quick-pick initiated by the plugins not displaying with a single item
Fixes #6050

- fixes issue where `quick-pick` which were initiated by the plugins was
not displayed when it only contains a single item. The fix is to include
a new flag `runIfSingle` which is used to determine if the quick-pick
should display or execute the item.
- The default value for `runIfSingle` is set to `true`.
- The `plugin-ext` sets the `runIfSingle` to `false` so it always displays
the quick-pick regardless if an single item is present. This behavior is aligned
with previous versions of Theia and VSCode.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>