yarn upgrade
Upgrade all our dependencies, according to our already specified ranges.

I had to tigthen some constraints because the latest version of some
packages cause issues. Notably, the typescript compiler has an issue on
3.6.3 but not 3.5.3, so I constrained it such as it wouldn't ever match
the version that causes issues.

In the long run, we should be more careful with the version ranges we
use, especially with runtime dependencies.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>