Fix formatting of 'common-frontend-contribution'
The following commit fixes the formatting from `common-frontend-contribution`.
There have been a large number of pull-requests that have updated the file and caused it
to have formatting issues (leading to noise during reviews). This commit addresses those
issues by fixing the formatting so it does not lead to unwanted changes during other pull-requests.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>