Cannot create a new file if the preferences widget is opened.
Fixes: https://github.com/eclipse-theia/theia/issues/7067 , https://github.com/eclipse-theia/theia/issues/7253

Initially we were unable to create a new file or folder while the user preferences was open.

The user-preferences has the file scheme set as `user-storage` and we did not fall back to the workspace root URI and instead were getting the URI from 'user_storage:settings.json'.
With this approach, we create files at certain directories and if a directory doesn't have the scheme set to 'file' it will fallback to the workspace root uri.

Signed-off-by: Muhammad Anas Shahid <muhammad.shahid@ericsson.com>