cli: improve plugin download logic
Fixes: #7501
Fixes: #7624

The following commit addresses issues with the `download:plugins`
script which is used to download plugins from various URLs (ex: 'open-vsx').

We do not wait for the request to finish before trying to write the
downloaded plugins on disk. requestretry also makes our life more
difficult by preventing us to easily use the response stream
asynchronously.

This commit changes the download logic so that it is more robust by
buffering the response stream and only replaying it to write on disk if
the request was successful.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>
Co-authored-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>