fix tree view - reveal not to invalidate item command
The issue is a regression introduced in https://github.com/eclipse-theia/theia/pull/8783
Calculation of parent chain (for id calculation) calls getChildren which dispose nodes.
This means that reveal for a node that is already revealed dispose the node and so the frontend has an id that doesn't exist in the backend anymore so tree item command is not found.

The fix is to check the cache before calling getChildren and making sure that the technical root is also in the cache.

Signed-off-by: Amiram Wingarten <amiram.wingarten@sap.com>