fix PreferenceChangeEvent<T> type
Despite precisely knowing for a given `preferenceName` what its
associated type for `newValue` and `oldValue` should be, the current
typing wasn't allowing TypeScript to properly infer the type.

With this commit, we can do things like:

    declare const event: PreferenceChangeEvent<{
        a: string
	b: number
    }>
    if (event.preferenceName === 'a') {
        event.newValue // type is `string`
    }

The type is more complex, but it now represents how it is meant to be
used rather than getting tsserver lost like before.

This commit also fixes places where the typing was broken and makes use
of the new information.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>