ci: run `dash-licenses` as part of GH workflow (#9953)
Add a "3PP FOSS license check" GitHub workflow that uses
the Eclipse Foundation dash-licences tool to validate the
licenses of our npm dependencies.

One can also run the license check locally with command:
"yarn license:check"

I have considered 3 alternatives:

1) Add the license check as a step of the build job of the
build workflow. The license check step needs to report
success or it will fail the build job and the publish job
(when it applies). If the license check reports success, it's
buried in the multiple steps of its job, not discoverable
from the PR page.

2) Add the license check as a new job in the build workflow.
This works better but there is redundancy in the job -  we need
to re-install node, checkout the repo, etc.

One advantage of this approach is that the license check has
visibility in the PR's CI checks section and can be made to fail
without impacting the other jobs.

3) _(chosen)_ Add a new workflow for the license check.

Like option 2 this gives the license check visibility in the PR's CI
checks section and the job can fail without impacting other
workflows/jobs.

This option has the additional benefit that it's quite generic
(for npm/yarn projects) and could be re-used more easily,
without changing the main workflow of the adopting repo.

---

It is worth noting that there are some dependencies that are
not yet completely cleared in this repository (which most likely
come from the time before Theia was an Eclipse project).

The tool exits with an error set because of those. The solution
implemented here is to have a "baseline" of dependencies
known to make `dash-licenses` trip, and to ignore the scan
error only if everything that is being reported is part of the
baseline.

This means that the GitHub Action check on PRs should only fail
if new "restricted" dependencies are introduced, which indicates
that actions must be taken, such as creating CQs.

Even when green it is worth checking the logs of the check since
it will report any dependency that is no longer required in the
baseline: This will happen once the various databases are
updated following the CQs we fill!

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>
Co-authored-by: Paul Marechal <paul.marechal@ericsson.com>