upgrade to `inversify@^5.1.1` (#9979)
Since we changed our TypeScript compilation target from ES5 to ES2017
there is a bug with `inversify` prior to version `5.1.1` leading to
the frontend application not starting.

We expected issues with `inversify` when targeting ES2017 but couldn't
find it in the end. This is because we are pulling a version of
`inversify` that did not cause the problem.

This commit updates the `inversify` range to pull a working version.