task: make `required` optional
The commit makes `TaskDefinition.properties.required` optional based on
the json schema validation definition. Previously, if an extension
defines a schema that omits the `required` property the application will
throw errors and fail to display the menu. The change treats the missing
`required` property as an empty array as per the spec.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>