navigator-decorator: add symlink icon and tooltip to linked files
This commit adds a new decorator for the file navigator. It displays a
downward right curved arrow at the tail of a file node if that file
contains a symbolic link. A tooltip is also added displaying "Symbolic
Link" when hovering over a symbolic linked file node's tail icon, and
will display the file's path only when hovering over the file itself.