dep: replace `colors` with `chalk`
The commit replaces the `colors.js` dependency with `chalk` due to a
recent deliberate breakage by the author. The `chalk` dependency is a
well maintained alternative by the open-source community.

The change also includes a migration guide update so that downstream
applications can include a resolution in their products if necessary.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>