Separate debug configuration providers from debug adapter contributors
The implementation of the vscode API for the resolution of initial
debug configurations as well as the resolution of specific
configurations was relying on the registered debug adapter
contributors rather than debug configuration providers,
this is causing issues e.g. #9978.

This implementation brings access to the debug configuration
providers all the way to the browser debug/plugin-debug-service,
where the providers can be filtered and then call its corresponding
API indirectly.

The former API is kept as deprecated.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>