fix: Omit _All Files_ filter in Electron on Linux (#11325)
* fix: Omit _All Files_ filter in Electron on Linux

We shouldn't add an _All Files_ filter, if we are on Linux and
no other filter is specified. Otherwise, we run into
https://github.com/eclipse-theia/theia/issues/11321
which causes an extension filter `['*']` to hide files without
extensions. Once we are on Electron >18, we should be able
to add the _All Files_ filter again in all cases, as this Electron
bug seems to be resolved in Electron above >18.

Fixes https://github.com/eclipse-theia/theia/issues/11321

Change-Id: Iec3120a6260f20b7f0d42c091aa00bde33fa915a
Signed-off-by: Philip Langer <planger@eclipsesource.com>

Co-authored-by: colin-grant-work <colin.grant@est.tech>